requirements.txt
fastapi
pymssql>=2.2.8
redis
types-redis
uvicorn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy
python-multipart
pytz
pydantic[email]
pandas
numpy
apscheduler
twilio
openpyxl
groq>=0.4.1
langchain>=0.1.0
langchain-groq>=0.0.1
python-dotenv>=1.0.0
phonenumbers==8.13.27
pdf2image==1.16.3
Pillow==10.1.0
aiofiles


